Refactor blocked actors into dedicated collection class
- Create new Blocked_Actors collection class for better code organization
- Move get_blocked_actors_with_count() and get_blocked_actors() methods from Moderation class
- Update all references to use the new collection methods
- Maintain consistent naming with other collection classes (Followers, Following)
- Improve separation of concerns by grouping blocked actor functionality
