+++ includes/wp-admin/table/class-blocked-actors.php
@@ -8,6 +8,7 @@
 namespace Activitypub\WP_Admin\Table;
 
 use Activitypub\Collection\Actors;
+use Activitypub\Collection\Blocked_Actors;
 use Activitypub\Moderation;
 use Activitypub\Sanitize;
 use Activitypub\Webfinger;
@@ -210,7 +211,7 @@ class Blocked_Actors extends \WP_List_Table {
 			$args['s'] = $this->normalize_search_term( \wp_unslash( $_GET['s'] ) ); // phpcs:ignore WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 		}
 
-		$blocked_with_count = Moderation::get_blocked_actors_with_count( $this->user_id, $per_page, $page_num, $args );
+		$blocked_with_count = Blocked_Actors::get_blocked_actors_with_count( $this->user_id, $per_page, $page_num, $args );
 
 		$blocked_actor_posts = $blocked_with_count['blocked_actors'];
 		$counter             = $blocked_with_count['total'];
